# some performance optimizations for starting apps