Some performance optimizations for starting apps:

- only seed the random value if we're going to use it (~600ms)
- switch to the sha256 implementation from `crypto` (~700ms => 70ms)
- default to not calling stop when starting an app. Android is ok w/ it, and stop isn't implemented for iOS (stop was ~500ms)
- reduce the number of times we call `simctl list --json` - each call was ~100ms, and we had 5-6 calls. Down to 2 now. This is the `isConnected` change. The semantics of that method are different now that we only create a `Device` instance for devices that actually exist