@devoncarew 2016-02-26
The biggest outstanding piece of work now is zipping the snapshot into an flx. We can re-write to shell out to `zip` when it's available, and save about 1000ms.